Options supports richtext, separated styles
