+++ src/styles/_components.scss
@@ -21,6 +21,7 @@
 @import './components/items';
 @import './components/list';
 @import './components/new-stage';
+@import './components/options';
 @import './components/overview';
 @import './components/panel';
 @import './components/protocol-stack';
